+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## Improvements:
 
+* Format Document will now right-align numeric values in tables. This can be overridden to left align them by setting "gherkin_table_cell_right_align_numeric_content = false" in .editorconfig file within a [*.feature] section. 
+
 ## Bug fixes:
+
 * Fix: Ambiguous steps reported wehn definition matches via more than one tag (#95)
+* Fix incorrect binding language detection for Visual Studio extension (#104)
+
 *Contributors of this release (in alphabetical order):* @clrudolphi
 
 # v2025.1.256 - 2025-03-07